refactor: switch to barrel exports
