+++ app/data/index.ts
@@ -1,0 +1,5 @@
+export * from "./internship";
+export * from "./navLinks";
+export * from "./projects";
+export * from "./techProducts";
+export * from "./urls";
